rag_backend/requirements.txt
python-multipart
uvicorn 
# Core
openai==1.57.0
environs==11.2.1
langchain==0.3.12

# Vector Store and Embeddings
# faiss-cpu==1.7.2  usar conda install -c pytorch faiss-cpu
faiss-cpu==1.7.2
sentence-transformers==3.3.1

# Api y Web
fastapi==0.115.6
uvicorn[standard]==0.20.0
python-multipart==0.0.9

# Document Processing
pypdf==5.1.0
unstructured==0.16.11

# Utils
tiktoken==0.8.0
requests==2.32.3
python-dotenv==1.0.0
numpy==1.24.3

# MongoDB
pymongo==4.10.1